+++ src/operators.ts
@@ -222,6 +222,7 @@ export function default_unknown_op(op: Bytes, args: SExp): Tuple<number, CLVMObj
     throw new EvalError("invalid operator", SExp.to(op));
   }
   
+  // The bytes here is 4bytes or smaller. So `int_from_bytes` is enough. (No bigint_from_bytes required)
   const cost_multiplier = int_from_bytes(op.subarray(0, op.length-1)) + 1;
   /*
     # 0 = constant